requirements.txt
numpy==1.23.5
pandas==1.5.3
fastapi==0.103.0
uvicorn==0.23.0
psycopg2-binary==2.9.8
python-dotenv==1.0.0
requests==2.31.0
